+++ homebytwo/importers/utils.py
@@ -1,9 +1,19 @@
+import csv
+from io import TextIOWrapper
+from tempfile import TemporaryFile
+from typing import Iterator
+from zipfile import ZipFile
+
+from django.contrib.gis.geos import Point
+from django.db import transaction
 from django.http import Http404
 
 from requests import Session, codes
 from requests.exceptions import ConnectionError
+from tqdm import tqdm
 
-from ..routes.models import Route
+from ..routes.models import Country, Place, PlaceType, Route
+from ..routes.models.place import PlaceTuple
 from .exceptions import SwitzerlandMobilityError, SwitzerlandMobilityMissingCredentials
 
 
@@ -98,3 +108,99 @@ def get_proxy_class_from_data_source(data_source):
         raise Http404("Data Source does not exist")
     else:
         return route_class
+
+
+def download_zip_file(url: str) -> ZipFile:
+    """
+    download zip file from remote url to bytes buffer
+    and wrap it in ZipFile
+    """
+
+    block_size = 1024
+    tmp_file = TemporaryFile()
+
+    with Session() as session:
+        response = session.get(url, stream=True)
+        response.raise_for_status()
+        file_size = int(response.headers["Content-Length"])
+
+        for data in tqdm(
+            response.iter_content(block_size),
+            total=int(file_size / block_size),
+            unit="B",
+            unit_scale=block_size,
+            desc=f"downloading from {url}",
+        ):
+            tmp_file.write(data)
+
+        return ZipFile(tmp_file)
+
+
+def get_csv_line_count(csv_file: TextIOWrapper, header: bool) -> int:
+    """
+    Get the number of features in the csv file
+    """
+    count = sum(1 for _ in csv.reader(csv_file))
+    csv_file.seek(0)  # return the pointer to the first line for reuse
+
+    return max(count - int(header), 0)
+
+
+def save_places_from_generator(
+    data: Iterator[PlaceTuple], count: int, source_info: str
+) -> str:
+    """
+    Save places from csv parsers in geonames.py or swissnames3d.py
+    """
+    created_counter = updated_counter = 0
+
+    with transaction.atomic():
+        for remote_place in tqdm(
+            data,
+            total=count,
+            unit="places",
+            unit_scale=True,
+            desc=f"saving places from {source_info}",
+        ):
+
+            # retrieve PlaceType from the database
+            try:
+                place_type = PlaceType.objects.get(code=remote_place.place_type)
+            except PlaceType.DoesNotExist:
+                print(f"Place type code: {remote_place.place_type} does not exist.")
+                continue
+
+            # country can be str or Country instance
+            country = remote_place.country
+            if country and not isinstance(country, Country):
+                try:
+                    country = Country.objects.get(iso2=remote_place.country)
+                except Country.DoesNotExist:
+                    print(f"Country code: {remote_place.country} does not exist.")
+                    continue
+
+            default_values = {
+                "name": remote_place.name,
+                "place_type": place_type,
+                "country": country,
+                "geom": Point(
+                    remote_place.longitude,
+                    remote_place.latitude,
+                    srid=remote_place.srid,
+                ),
+                "altitude": remote_place.altitude,
+            }
+
+            # create or update Place
+            _, created = Place.objects.update_or_create(
+                data_source=remote_place.data_source,
+                source_id=remote_place.source_id,
+                defaults=default_values,
+            )
+
+            created_counter += int(created)
+            updated_counter += int(not created)
+
+        return "Created {} new places and updated {} places. ".format(
+            created_counter, updated_counter
+        )